rename shared type file to be more explicit
